% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidysql.R
\docType{package}
\name{tidysql-package}
\alias{tidysql}
\alias{tidysql-package}
\title{tidysql: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mgirlich/tidysql}
  \item Report bugs at \url{https://github.com/mgirlich/tidysql/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Girlich \email{maximilian.girlich@metoda.com}

}
\keyword{internal}
